added files for interfaces and types
