+++ src/lib/formatNumber.ts
