+++ src/stores/budget/budget.types.ts
@@ -1,0 +1,17 @@
+export type BudgetPoolType =
+  | "personal"
+  | "bills"
+  | "travel"
+  | "food"
+  | "savings";
+
+export interface IBudgetPool {
+  type: BudgetPoolType;
+  label: string;
+  percentage: number;
+}
+
+export interface IMonthlyBudget {
+  month: string;
+  pools: IBudgetPool[];
+}
