+++ src/stores/members/members.types.ts
@@ -1,0 +1,5 @@
+export interface IMember {
+  id: string;
+  name: string;
+  status: "active" | "inactive";
+}
