+++ src/stores/monthlyIncome/monthlyIncome.types.ts
@@ -1,0 +1,9 @@
+import { BudgetPoolType } from "../budget/budget.types";
+
+export interface IMonthlyIncome {
+  id: string;
+  memberId: string;
+  month: string;
+  salary: number;
+  breakdown: Record<BudgetPoolType, number>;
+}
